--- shoppingMall/src/pages/SignUp.module.css
@@ -1,107 +1,0 @@
-* {
-  box-sizing: border-box;
-}
-
-button {
-  cursor: pointer;
-}
-
-.container {
-  background: #f9f9f9;
-  display: flex;
-  justify-content: center;
-  padding: 80px;
-}
-
-.form-container {
-  width: 700px;
-  padding: 10px 30px;
-  background: white;
-  box-shadow: rgba(0, 0, 0, 0.1) 0px 0px 15px;
-}
-
-.content {
-  text-align: center;
-  border-bottom: 2px solid #1d1d1d;
-}
-
-.content h2 {
-  font-size: 30px;
-}
-
-.sign-up-form {
-  display: flex;
-  flex-direction: column;
-  gap: 25px;
-  padding: 50px;
-}
-
-.input-container {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.input-container:nth-child(2) input {
-  width: 302px;
-  margin-left: auto;
-  margin-right: 10px;
-}
-
-.input-container input {
-  padding: 10px;
-  width: 400px;
-  outline: none;
-  border: 1px solid #dcdbdc;
-}
-
-.input-container:nth-child(7) input {
-  border: none;
-  padding: 0;
-}
-
-.duplicate-check-button,
-input[type="file"]::-webkit-file-upload-button {
-  height: 37.5px;
-  padding: 0px 20px;
-  background: #565656;
-  color: white;
-  border: 1px solid #565656;
-  cursor: pointer;
-}
-
-.duplicate-check-button:hover,
-input[type="file"]::-webkit-file-upload-button:hover {
-  background: #1c1c1c;
-}
-
-.button-container {
-  text-align: center;
-  margin-top: 25px;
-}
-
-.button-container button {
-  width: 250px;
-  padding: 11px;
-  font-weight: bold;
-}
-
-.cancel-button {
-  border: 1px solid #565656;
-  background: white;
-}
-
-.cancel-button:hover {
-  background: #f5f5f5;
-}
-
-.confirm-button {
-  border: 1px solid #3471f0;
-  background: #3471f0;
-  color: white;
-  margin-left: 10px;
-}
-
-.confirm-button:hover {
-  background: #234896;
-}
